# src/api.yaml
openapi: 3.1.0
info:
  title: TrustRAG Studio API
  version: 1.0.0
  description: |
    TrustRAG Studio 后端接口文档（FastAPI）。
    包含两个主要模块：
    1. **Corpus (知识库)**: 文件上传、管理、切片配置与向量索引构建。
    2. **Chat (对话)**: 支持 RAG/纯 LLM 模式、模型选择、Web 搜索及可解释性（Query Rewrite/Claim Check）的流式对话。
servers:
  - url: http://localhost:8000/api/v1
    description: Local environment
...
paths:
...
  /knowledge/upload:
    post:
      tags: [KnowledgeBase]
      operationId: uploadFile
      summary: 上传文件至知识库
      description: 对应图1右下角 "Upload a knowledge file"。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                knowledge_name:
                  type: string
                  default: "default_index"
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploaded"
...
components:
  schemas:
...
    FileUploaded:
      type: object
      properties:
        success: { type: boolean, example: true }
        filename: { type: string }
        file_id: { type: string }

    # --- Chat Models ---
...